@@ -1,19 +1,15 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Builder;
+﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.HttpsPolicy;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Sample.Mvc.Models;
+using Microsoft.Extensions.Hosting;
+using Raven.Client.Documents;
 using Raven.DependencyInjection;
 using Raven.Identity;
-using Raven.Client.Documents;
 using Sample.Mvc.Common;
+using Sample.Mvc.Models;
+using System.Collections.Generic;
 
 namespace Sample.Mvc
 {
@@ -41,12 +37,12 @@ namespace Sample.Mvc
                 .AddRavenDbDocStore() // Create our IDocumentStore singleton using the database settings in appsettings.json
                 .AddRavenDbAsyncSession() // Create an Raven IAsyncDocumentSession for every request.
                 .AddRavenDbIdentity<AppUser>(); // Let Raven store users and roles.
-            
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+
+            services.AddControllersWithViews();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
@@ -59,22 +55,24 @@ namespace Sample.Mvc
                 app.UseHsts();
             }
 
-            app.UseAuthentication();
             app.UseHttpsRedirection();
+            app.UseRouting();
             app.UseStaticFiles();
             app.UseCookiePolicy();
+
+            // Required for Raven Identity to work with authorization and authentication.
+            app.UseAuthentication();
+            app.UseAuthorization();
 
             // Create the database if it doesn't exist.
             // Also, create our roles if they don't exist. Needed because we're doing some role-based auth in this demo.
             var docStore = app.ApplicationServices.GetRequiredService<IDocumentStore>();
             docStore.EnsureExists();
             docStore.EnsureRolesExist(new List<string> { AppUser.AdminRole, AppUser.ManagerRole });
-            
-            app.UseMvc(routes =>
+
+            app.UseEndpoints(endpoints =>
             {
-                routes.MapRoute(
-                    name: "default",
-                    template: "{controller=Home}/{action=Index}/{id?}");
+                endpoints.MapControllerRoute("default", "{controller=Home}/{action=Index}/{id?}");
             });
         }
     }
